+++ tests/integration/pipeline.ts
@@ -198,7 +198,7 @@ describe('PipelineEditor', () => {
   it('runs invalid pipeline', () => {
     cy.get('.run-action button').click();
     cy.get('.MuiAlert-message').should('be.visible');
-    cy.get('image[data-id="node_dec_image_2_error"]').should('exist');
+    cy.get('.d3-node-dec-image').should('exist');
 
     // closes alert message
     cy.get('.MuiAlert-action > button[aria-label="close"]').click();
@@ -208,7 +208,7 @@ describe('PipelineEditor', () => {
     cy.wait(1000);
     cy.get('.export-action button').click();
     cy.get('.MuiAlert-message').should('be.visible');
-    cy.get('image[data-id="node_dec_image_2_error"]').should('exist');
+    cy.get('.d3-node-dec-image').should('exist');
 
     // closes alert message
     cy.get('.MuiAlert-action > button[aria-label="close"]').click();